handlers/requirements.txt
aiogram>=3.26.0
aiohttp==3.10.10
aiofiles==24.1.0
aiosqlite==0.20.0
asyncpg==0.29.0
Pillow==11.0.0
python-dotenv==1.0.1
httpx==0.27.2
